requirements.txt
RPi.GPIO
smbus2
requests
jsonschema
